requirements.txt
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
defusedxml==0.5.0
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
djongo==1.2.32
enum34==1.1.6
idna==2.8
inflection==0.3.1
oauthlib==3.0.1
psycopg2==2.8.1
PyJWT==1.7.1
pymongo==3.5.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.2.4
urllib3==1.24.1
vk==2.0.2
vk-api==11.4.0
